Stop the flush window's estimator from being dragged up by rare slow syncs
The adaptive window is set to ceil(sync_us_ewma) capped by the configured ceiling, so the estimator is
the policy.  Its creep-up path (+= (dur-ewma)/8) let one multi-millisecond sync add ~600 us, and the
guest's window therefore sat at 2-3 ms on a device whose typical syncs are 0.2-0.7 ms.  Measured cost:
67247 ops/s at a pinned 1 ms window against 35822 with the inflated one, p99 8898 us against 17072 us -
the mechanism confirmed by window_ms itself, not by comparing runs from different minutes.  A sample
more than 4x the current estimate no longer raises it; outliers are still recorded in sync_us_max and
counted in slow_syncs.  After the change the guest converges to window_ms=1 and reads 57247 ops/s with
p99 8886 us at the same configured ceiling, and Windows still adapts upward (window_ms=2 at a 1.1 ms
sync) where 3 ms measured best.

Also fixes the three -Wpointer-sign diagnostics the pinned build gate caught in the new init options
(k_parse_uint writes an int, not an unsigned int).

+++ code/kdbsvr.c
@@ -387,7 +387,7 @@ static int k_run_server_args(int argc,char **argv){
 }
 static int k_run_init_args(int argc,char **argv){
   k_cfg cfg;
-  unsigned int value;
+  int value;   /* k_parse_uint writes an int; using unsigned int here tripped -Wpointer-sign */
   int i;
   if(argc<3||k_cfg_reset(argv[2])!=0||k_wal_meta_init(argv[2])!=0) return -1;
   if(argc==3){
@@ -405,14 +405,14 @@ static int k_run_init_args(int argc,char **argv){
       return -1;
     }
     if(strcmp(argv[i],"--flush-items")==0){
-      if(k_parse_uint(argv[i+1],65536,&value)!=0||value<1u){ printf("kdbsvr: fatal: --flush-items must be 1..65536\n"); return -1; }
-      cfg.flush_item_limit=value;
+      if(k_parse_uint(argv[i+1],65536,&value)!=0||value<1){ printf("kdbsvr: fatal: --flush-items must be 1..65536\n"); return -1; }
+      cfg.flush_item_limit=(unsigned int)value;
     }else if(strcmp(argv[i],"--flush-window-ms")==0){
-      if(k_parse_uint(argv[i+1],60000,&value)!=0||value<1u){ printf("kdbsvr: fatal: --flush-window-ms must be 1..60000\n"); return -1; }
-      cfg.flush_timeout_ms=value;
+      if(k_parse_uint(argv[i+1],60000,&value)!=0||value<1){ printf("kdbsvr: fatal: --flush-window-ms must be 1..60000\n"); return -1; }
+      cfg.flush_timeout_ms=(unsigned int)value;
     }else if(strcmp(argv[i],"--flush-bytes")==0){
-      if(k_parse_uint(argv[i+1],268435456,&value)!=0||value<1u){ printf("kdbsvr: fatal: --flush-bytes must be 1..268435456\n"); return -1; }
-      cfg.flush_bytes_limit=value;
+      if(k_parse_uint(argv[i+1],268435456,&value)!=0||value<1){ printf("kdbsvr: fatal: --flush-bytes must be 1..268435456\n"); return -1; }
+      cfg.flush_bytes_limit=(unsigned int)value;
     }else{
       printf("kdbsvr: fatal: unknown init option '%s'\n",argv[i]);
       return -1;
